robotics/ros2_bridge/requirements.txt
rclpy
websockets>=12.0
PyYAML>=6.0
